fix: credits возвращается как int, добавила yookassa в зависимости
- PaymentCreateResponse.credits изменён с Decimal на int — Pydantic сериализовал
  Decimal как строку, что ломало тест и клиент на JS
- Явное приведение int(credits) в endpoint create_payment
- Добавила yookassa==3.1.0 в requirements.txt (пропустила в предыдущем коммите)
- Все 70 тестов проходят, покрытие 70%

+++ backend/app/schemas/billing.py
@@ -45,7 +45,7 @@ class PaymentCreateResponse(BaseModel):
     payment_id: int
     yookassa_id: str
     confirmation_url: str
-    credits: Decimal
+    credits: int
     amount_rub: int
     test_mode: bool = False
